Drop Travis as our CI provider
Since travis changed its pricing policy it has become more limited in
what we can accomplish with it. It now uses a limited credit model for
open source projects which besides being cumbersome puts the project in
a precarious position where we may be unable to run tests if we dont
have anymore credits.

For this reason we will be moving to github actions since they seem to
be the best alternative at the moment