Add --ignore-errors option
This commit adds a new option --ignore-errors. Originally it's a
variable defined in get_all_imports and is always False. The purpose is
to ignore errors when parsing single .py file. Also the opening/reading
file part is moved to the try block, as those operations may also
fail(encoding problem).

Signed-off-by: 司芳源 <sify1221@gmail.com>